Dismiss notification when opening dialog
