+++ packages/edit-site/src/components/page-content-focus-manager/index.js
@@ -8,7 +8,6 @@ import { useSelect } from '@wordpress/data';
  */
 import { store as editSiteStore } from '../../store';
 import DisableNonPageContentBlocks from './disable-non-page-content-blocks';
-import EditTemplateDialog from './edit-template-dialog';
 import EditTemplateNotification from './edit-template-notification';
 import BackToPageNotification from './back-to-page-notification';
 
@@ -19,12 +18,7 @@ export default function PageContentFocusManager( { contentRef } ) {
 	);
 	return (
 		<>
-			{ hasPageContentFocus && (
-				<>
-					<DisableNonPageContentBlocks />
-					<EditTemplateDialog contentRef={ contentRef } />
-				</>
-			) }
+			{ hasPageContentFocus && <DisableNonPageContentBlocks /> }
 			<EditTemplateNotification contentRef={ contentRef } />
 			<BackToPageNotification />
 		</>
